selinux: ignore invalid contexts in kernfs_init_security in child namespace
kernfs_init_security() tries to propagate the parent directory's
SELinux xattr to the child (or compute a transition, depending on
policy) when a child node is created. In the case of cgroup2,
this can cause a hard failure when trying to create any files
in the cgroup2 mount for a child SELinux namespace if the
context in the parent is invalid in the child. Treat this
the same way as if the parent lacked a SELinux xattr and just
let the child be labeled based on policy.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>